@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class SectionTitle extends StatelessWidget {
-  const SectionTitle({Key key, this.title}) : super(key: key);
+  const SectionTitle({Key? key, required this.title}) : super(key: key);
 
   final String title;
 
